Update metrics name
